+++ deser-value/src/ser.rs
@@ -67,6 +67,7 @@ impl Serialize for Seq {
         ContainerShape::new()
             .with_len(self.len())
             .with_order(self.order())
+            .with_repeated(self.is_repeated())
     }
 }